ajout de la gestion de fichiers

+++ Controllers/Books.js
@@ -1,31 +1,67 @@
 const Book = require('../Models/Book');
+const fs = require('fs');
 
 exports.createBook = (req, res, next) => {
-    const book = new Book({
-      ...req.body // Insère toutes les données du corps de la requête
-    });
-    book.save()
-      .then(() => res.status(201).json({ message: 'Livre enregistré !' }))
-      .catch(error => res.status(400).json({ error }));
-  };
+  const bookObject = JSON.parse(req.body.book);
+  delete bookObject._id;
+  delete bookObject._userId;
+  
+  const book = new Book({
+    ...bookObject,
+    userID: req.auth.userID,
+    imageUrl: `${req.protocol}://${req.get('host')}/Images/${req.file.filename}`
+  });
+  book.save()
+    .then(() => res.status(201).json({ message: 'Livre enregistré !' }))
+    .catch(error => res.status(400).json({ error }));
+};
 
   exports.getOneBook = (req, res, next) => {
     Book.findOne({ _id: req.params.id })
       .then(book => res.status(200).json(book))
       .catch(error => res.status(404).json({ error }));
-  };
+};
 
   exports.modifyBook = (req, res, next) => {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
+    const bookObject = req.file ? {
+      ...JSON.parse(req.body.book),
+      imageUrl: `${req.protocol}://${req.get('host')}/Images/${req.file.filename}`
+    } : { ...req.body };
+    
+    delete bookObject._userId;
+    Book.findOne({_id: req.params.id})
+      .then((book) => {
+        if (book.userId != req.auth.userId) {
+          res.status(401).json({ message : 'Not authorized'});
+      } else {
+        Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
       .then(() => res.status(200).json({ message: 'Livre modifié !'}))
-      .catch(error => res.status(400).json({ error }));
-  };
+      .catch(error => res.status(401).json({ error }));
+        }
+      })
+      .catch((error) => {
+        res.status(400).json({ error });
+      })
+    };
 
-  exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Livre supprimé !'}))
-      .catch(error => res.status(400).json({ error }));
-  };
+exports.deleteBook = (req, res, next) => {
+    Book.findOne({ _id: req.params.id})
+      .then(book => {
+        if (book.userId != req.auth.userId) {
+          res.status(401).json({message: 'Not Authorized'});
+        } else {
+          const filename = book.imageUrl.split('/Images/')[1];
+          fs.unlink(`Images/${filename}`, ()=> {
+            Book.deleteOne({ _id: req.params.id })
+            .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+            .catch(error => res.status(401).json({ error }));
+            });
+        }
+      })
+    .catch( error => {
+      res.status(500).json({ error });
+    });
+};
 
   exports.getAllBooks = (req, res, next) => {
     console.log('Requête GET reçue');
